lms/requirements.txt
asgiref==3.7.2                                                                                      
black==22.12.0                                                                                      
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
Deprecated==1.2.14
Django==4.0.8
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.1.3
django-filter==23.5
django-jet-reboot==1.3.5
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gopay==2.0.1
idna==3.6
mypy-extensions==1.0.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.5
pydantic==1.10.14
PyJWT==2.8.0
python-decouple==3.8
pytz==2022.7
reportlab==4.0.4
requests==2.31.0
setuptools==69.1.1
sqlparse==0.4.4
stripe==5.5.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.2.0
wrapt==1.16.0

drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
# pytz==2024.1
PyYAML==6.0.1
uritemplate==4.1.1